requirements.txt
flask
pillow
numpy
waitress